pyproject.toml
[project]
name = "graphrefly"
version = "0.7.0"
description = "Reactive graph protocol for human + LLM co-operation. Composable nodes, glitch-free diamond resolution, two-phase push, durable streaming. Zero dependencies."
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [{ name = "David Chen" }]
keywords = [
	"reactive",
	"graph",
	"automation",
	"workflow",
	"llm",
	"ai-agents",
	"orchestration",
	"state-management",
	"signals",
	"streaming",
	"observable",
	"causal-tracing",
	"human-in-the-loop",
	"natural-language",
	"durable-workflow",
	"asyncio",
	"zero-dependency",
]
classifiers = [
	"Development Status :: 3 - Alpha",
	"Intended Audience :: Developers",
	"License :: OSI Approved :: MIT License",
	"Programming Language :: Python :: 3",
	"Programming Language :: Python :: 3.12",
	"Programming Language :: Python :: 3.13",
	"Programming Language :: Python :: 3.14",
	"Topic :: Software Development :: Libraries :: Python Modules",
	"Topic :: Scientific/Engineering :: Artificial Intelligence",
	"Framework :: AsyncIO",
	"Typing :: Typed",
]

[project.urls]
Homepage = "https://py.graphrefly.dev"
Repository = "https://github.com/graphrefly/graphrefly-py"
Documentation = "https://py.graphrefly.dev"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/graphrefly"]

[dependency-groups]
dev = [
	"pytest>=8.0",
	"pytest-asyncio>=0.24",
	"pytest-benchmark>=5.0",
	"python-semantic-release>=10.5.3,<11",
	"ruff>=0.9",
	"mypy>=1.14",
	"croniter>=2.0",
	"django>=4.2",
	"fastapi>=0.100",
	"httpx>=0.24",
	"pillow>=12.2.0",
	"trio>=0.33.0",
]

[tool.semantic_release]
version_toml = ["pyproject.toml:project.version"]
build_command = "uv build"
commit_message = "chore(release): {version} [skip ci]\n\nAutomatically generated by python-semantic-release"
# TODO: Remove this once we have a proper release process.
allow_zero_version = true

[tool.ruff]
line-length = 100
indent-width = 4
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "TC"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-m 'not benchmark'"
markers = [
	"benchmark: micro-benchmarks (run: pytest -m benchmark tests/bench_core.py --benchmark-only)",
]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
